Move caching to dedicated sub-project.
