+++ caching/build.sbt
@@ -1,0 +1,11 @@
+import Dependencies._
+
+name := "common-caching"
+
+libraryDependencies ++= Seq(
+  jedis,
+  mockJedis,
+  scalaGuice,
+  sprayJson,
+  typesafeConfig
+)
