+++ caching/src/test/scala/org/allenai/common/cache/JsonQueryCacheSpec.scala
