+++ core/build.sbt
@@ -4,11 +4,9 @@ name := "common-core"
 
 libraryDependencies ++= Seq(
   apacheLang3,
-  jedis,
   Logging.logbackClassic,
   Logging.logbackCore,
   Logging.slf4jApi,
-  mockJedis,
   sprayJson,
   typesafeConfig
 )
